Data migration optimization

+++ webclient/adb/models.py
@@ -24,10 +24,10 @@ class Kladysm5(ExportModelOperationsMixin("kladysm5"), models.Model):
     """
 
     gid = models.IntegerField(primary_key=True)
-    id = models.DecimalField(max_digits=1000, decimal_places=500)
+    id = models.IntegerField()
     mapname = models.TextField()
     mapno = models.TextField()
-    podil = models.DecimalField(max_digits=1000, decimal_places=999)
+    podil = models.DecimalField(max_digits=10, decimal_places=9)
     geom = pgmodels.PolygonField(srid=5514)
     cislo = models.TextField()
